chore: move properties to files in new folders, create index files, and update imports
