+++ dashboard/src/components/index.ts
@@ -1,0 +1,7 @@
+// src/components/index.ts
+export { default as CreateItemModal } from './CreateItemModal';
+export { default as EditItemModal } from './EditItemModal';
+export { default as DownloadButton } from './DownloadButton';
+export { default as Footer } from './Footer';
+export { default as Header } from './Header';
+export { Modal } from './Modal';
