+++ dashboard/src/constants/fields.ts
@@ -1,0 +1,9 @@
+import { Field } from "@/types/";
+
+export const fields: Field[] = [
+  { label: "SKU", value: "item_sku" },
+  { label: "Name", value: "item_name" },
+  { label: "Color", value: "item_color" },
+  { label: "Size", value: "item_size" },
+  { label: "Quantity", value: "item_quantity" },
+];
