+++ dashboard/src/types/field.ts
@@ -1,0 +1,5 @@
+// src/types/item.ts
+export interface Field {
+  label: string;
+  value: string;
+}
