+++ dashboard/src/types/index.ts
@@ -1,0 +1,3 @@
+// src/types/index.ts
+export * from './item'
+export * from './field';
